requirements.txt
alembic==1.7.3
certifi==2021.5.30
charset-normalizer==2.0.6
docker==5.0.2
greenlet==1.1.1
idna==3.2
loguru==0.5.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
schedule==1.1.0
SQLAlchemy==1.4.25
typeguard==2.12.1
urllib3==1.26.7
websocket-client==1.2.1
